Improve the human AI.
Added a second human to shoot at.
Experimented with creating a child Human_Patrol class (called Human_Patrol_Path_A).  This is the patrol path for the second human AI.  Nothing fancy but glad it works that way.
Added a Health int, and an IsDead bool to the Human pawn.
Added functions that can be called for body shots and headshots.